requirements.txt
blinker==1.5
click==8.1.3
colorama==0.4.6
defusedxml==0.7.1
Flask==2.2.2
Flask-Mail==0.9.1
fonttools==4.38.0
fpdf2==2.5.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
python-dateutil==2.8.2
pytz==2022.6
six==1.16.0
svg.path==6.2
Werkzeug==2.2.2
